requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

# LLM & AI
groq==0.11.0
langchain==0.3.7
langchain-community==0.3.5
langchain-groq==0.2.1
sentence-transformers==3.2.1

# Database
psycopg2-binary==2.9.10
pgvector==0.3.5
sqlalchemy<2.0.36

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Utilities
httpx==0.27.2
aiofiles==24.1.0
pypdf==5.1.0
python-docx==1.1.2
markdown==3.7

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
